몬스터 죽는 상태 설정
타입이 맞지 않는 값들 형변환

디버그,릴리즈 모드의 플래그 설정

+++ DX_Project/DX_Framework.cpp
@@ -186,8 +186,8 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
         break;
     case WM_MOUSEMOVE:
     {
-        mouse_pos.x = GET_X_LPARAM(lParam);
-        mouse_pos.y = GET_Y_LPARAM(lParam);
+        mouse_pos.x = (float)GET_X_LPARAM(lParam);
+        mouse_pos.y = (float)GET_Y_LPARAM(lParam);
         break;
     }
     case WM_DESTROY:
