+++ DX_Project/Scenes/MapleIsland.h
@@ -14,7 +14,7 @@ private:
 	
 	
 public:
-	MapleIsland(float area);
+	MapleIsland(UINT area);
 	~MapleIsland();
 
 	// Scene을(를) 통해 상속됨
