+++ DX_Project/Scenes/RunningWaitting.cpp
@@ -1,6 +1,6 @@
 #include "framework.h"
 
-RunningWaitting::RunningWaitting(float area)
+RunningWaitting::RunningWaitting(UINT area)
 {
 	bg = new BackGroundUV(L"Texture/Image/RunningWait.png", Vector2(0, 0), Vector2(1, 1), Vector2(WIN_CENTER_X, WIN_CENTER_Y), Vector2(3, 3), 0.0f, Vector2(SCREEN_SIZE_X, SCREEN_SIZE_Y));
 	player = new AnimatePlayer(L"Texture/AnimateScene/Animation/sprite.png");
