+++ DX_Project/Scenes/StarBridgeWaitting.cpp
@@ -1,6 +1,6 @@
 #include "framework.h"
 
-StarBridgeWaitting::StarBridgeWaitting(float area)
+StarBridgeWaitting::StarBridgeWaitting(UINT area)
 {
 	bg = new BackGroundUV(L"Texture/Image/starbridgewait.png", Vector2(0, 0), Vector2(1, 1), Vector2(WIN_CENTER_X, WIN_CENTER_Y), Vector2(3, 3), 0.0f, Vector2(SCREEN_SIZE_X, SCREEN_SIZE_Y));
